add ollama resource request/limit
Added resource request to make sure it gets on a node with a reasonable
amount of resources for it to respond within a minute for smaller
models. Added limits because some models took the entire node's memory
and stalled it momentarily until k8s killed the container.

Also, disabled the nvidia specification until a node with the `nvidia`
resource is actually available.