fix csi-driver-nfs pod check and add nfs-delete StorageClass
The csi-driver-nfs pod check would always pass because it would negative
match on the kubectl header. Fixed to not print the header and fail if
any pod wasn't in the Running state.

Add the `nfs-delete` StorageClass so that deployments would use that for
dynamically provisioned PVs instead of retaining unwanted PVs. PVs that
should be retained (and reused by a Pod, presumably) would use the
`nfs-retain` StorageClass in tandem with a PV definition to ensure it is
bound to a Pod's PVC.